New experimental feature: stage templates
This allows to factor out boiler-plate pipeline bits into separate files.  Then
pipelines can define a `templates` section to inform logbus where to load them
from.  Then pipeline stages can provide `template` to merge in that template's
definition in with the stage.

The implementation is janky, mostly a copy-paste of how plugins are loaded.  The
index.js is a poor example of how to write code, but it works.